feat: Upgraded extension to use workspaceState storage
Upgraded the extension to use the workspaceState Memento storage exposed through the ExtensionContext to store aliases. instead of storing them as configuration in the settings.json files.

BREAKING CHANGE: Updating the method used to store aliases is a breaking change that will render any existing aliases configured in the settings.json file non-functional. Existing aliases will need to be manually re-created when upgrading to the new version.